create abstract models
