+++ cities_light/contrib/autocompletes.py
@@ -1,6 +1,10 @@
-from ..models import Country, Region, City
+from ..loading import get_cities_model
 
 import autocomplete_light
+
+Country = get_cities_model('Country')
+Region = get_cities_model('Region')
+City = get_cities_model('City')
 
 
 class CityAutocomplete(autocomplete_light.AutocompleteModelBase):
