+++ cities_light/contrib/restframework.py
@@ -29,7 +29,11 @@ from djangorestframework.views import ModelView, ListModelView
 from djangorestframework.mixins import InstanceMixin, ReadModelMixin
 from djangorestframework.resources import ModelResource
 
-from ..models import Country, Region, City
+from ..loading import get_cities_model
+
+Country = get_cities_model('Country')
+Region = get_cities_model('Region')
+City = get_cities_model('City')
 
 
 class CityResource(ModelResource):
