+++ cities_light/loading.py
@@ -1,0 +1,17 @@
+import django
+from .settings import CITIES_LIGHT_APP_NAME
+
+
+if django.VERSION < (1, 7):
+    from django.db.models import get_model
+else:
+    from django.apps import apps
+    get_model = apps.get_model
+
+
+def get_cities_model(model_name, *args, **kwargs):
+    """
+    Returns cities model, either default either customised, depending on
+    ``settings.CITIES_LIGHT_APP_NAME``.
+    """
+    return get_model(CITIES_LIGHT_APP_NAME, model_name, *args, **kwargs)
